worker/requirements.txt
amqp==2.6.1
asgiref==3.7.2
attrs==23.1.0
billiard==3.6.4.0
celery==4.4.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.6
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-spectacular==0.26.5
future==1.0.0
idna==3.7
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kombu==4.6.11
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rpds-py==0.13.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.1
vine==1.3.0
wcwidth==0.2.13
